Relocate interface declarations into subpackage
